+++ sdk/communication/communication-chat/src/models/events.ts
@@ -1,0 +1,30 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import {
+  ChatEventId,
+  ChatMessageReceivedEvent,
+  ChatMessageEditedEvent,
+  ChatMessageDeletedEvent,
+  ReadReceiptReceivedEvent,
+  TypingIndicatorReceivedEvent,
+  ChatThreadCreatedEvent,
+  ChatThreadDeletedEvent,
+  ChatThreadPropertiesUpdatedEvent,
+  ParticipantsAddedEvent,
+  ParticipantsRemovedEvent
+} from "@azure/communication-signaling";
+
+export {
+  ChatEventId,
+  ChatMessageReceivedEvent,
+  ChatMessageEditedEvent,
+  ChatMessageDeletedEvent,
+  ReadReceiptReceivedEvent,
+  TypingIndicatorReceivedEvent,
+  ChatThreadCreatedEvent,
+  ChatThreadDeletedEvent,
+  ChatThreadPropertiesUpdatedEvent,
+  ParticipantsAddedEvent,
+  ParticipantsRemovedEvent
+};