+++ sdk/redis/arm-rediscache/src/redisManagementClientContext.ts
@@ -43,7 +43,7 @@ export class RedisManagementClientContext extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-rediscache/30.0.0-beta.1`;
+    const packageDetails = `azsdk-js-arm-rediscache/6.0.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`