+++ sdk/identity/identity/CHANGELOG.md
@@ -110,6 +110,8 @@ Azure Service Fabric support hasn't been added on the initial version 2 of Ident
 - Renamed the `ApplicationCredential` to `AzureApplicationCredential`.
 - Removed the `CredentialPersistenceOptions` from `DefaultAzureCredential` and `EnvironmentCredential`.
 - Merged the configuration and the options bag on the `OnBehalfOfCredential` into a single options bag.
+- `AuthenticationRequiredError` (introduced in 2.0.0-beta.1) now has its parameters into a single options bag.
+- `AuthenticationRequiredError` (introduced in 2.0.0-beta.1) now has its parameters in a single options bag, `AuthenticationRequiredErrorOptions`.
 
 ### Bugs Fixed